requirements.txt
rdkit==2022.3.5
mysql-connector-python==8.0.31
requests==2.28.1
scipy==1.10.0
matplotlib==3.5.3